move trend indicators to separate file
trend indicators considered to be indicators where signals aren't based
on crossing a upper/lower/zero line. still need to adjust for whether
we want to separate volatility indicators.
