+++ src/lib.rs
@@ -2,6 +2,7 @@
 //!
 //! a set of functions to derive signals from financial(-like) data
 
-pub mod indicator;
+pub mod momentum;
 pub mod smooth;
+pub mod trend;
 pub mod volume;
